# photo360-common/src/main/java/pl/put/photo360/dto/LoginRequestDto.java
package pl.put.photo360.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

/**
 * Login request dto, which contains all required data to authenticate the user.
 */
@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class LoginRequestDto
{
    private String login;
    private String password;
}